+++ src/plonemeeting/portal/core/faceted/criteria.py
@@ -4,6 +4,7 @@ from eea.facetednavigation.criteria.handler import Criteria
 from plone import api
 from plonemeeting.portal.core.config import CONFIG_FOLDER_ID
 from plonemeeting.portal.core.config import FACETED_FOLDER_ID
+from plonemeeting.portal.core.config import FACETED_MANAGER_CRITERIA
 
 import copy
 
@@ -23,4 +24,8 @@ class MeetingsCriteria(Criteria):
         faceted = getattr(config_folder, FACETED_FOLDER_ID)
         self.context = faceted
         criteria = copy.deepcopy(self._criteria())
+        # remove FACETED_MANAGER_CRITERIA if current user is anonymous
+        if api.user.is_anonymous():
+            criteria = [criterion for criterion in criteria
+                        if criterion.__name__ not in FACETED_MANAGER_CRITERIA]
         return criteria