+++ src/plonemeeting/portal/core/tests/test_meeting.py
@@ -5,7 +5,8 @@ from plonemeeting.portal.core.tests.portal_test_case import PmPortalDemoFunction
 class TestMeetingView(PmPortalDemoFunctionalTestCase):
 
     def test_call_meeting_view_as_manager(self):
-        meeting = self.portal["belleville"].listFolderContents({"portal_type": "Meeting"})[0]
+        meeting = self.portal["belleville"].listFolderContents(
+            {"portal_type": "Meeting"})[0]
         self.login_as_manager()
         request = self.portal.REQUEST
         view = meeting.restrictedTraverse("@@view")
@@ -14,10 +15,13 @@ class TestMeetingView(PmPortalDemoFunctionalTestCase):
         self.assertEqual(view.request.response.status, 200)
 
     def test_call_meeting_view_as_anonymous(self):
-        meeting = self.portal["belleville"].listFolderContents({"portal_type": "Meeting"})[-1]
+        meeting = self.portal["belleville"].listFolderContents(
+            {"portal_type": "Meeting"})[-1]
         self.login_as_test()
         view = meeting.restrictedTraverse("@@view")
         view()
         self.assertEqual(view.request.response.status, 302)
-        self.assertDictEqual(view.request.response.headers,
-                             {'location': 'http://nohost/plone/belleville/meetings#seance={}'.format(meeting.UID())})
+        self.assertDictEqual(
+            view.request.response.headers,
+            {'location': 'http://nohost/plone/belleville/seances#seance={}'.format(
+                meeting.UID())})